Removed unused ViewController import.
This import originally was importing MainViewController.h, which is Cordova's view controller.
It was updated to use ViewController.h as part of our self-managed cordova workflow, which
ended up using XCode generated view controller name instead of MainViewController.

This created an incompatilibility between the two environments, but this import is not used anyway
so we can simply remove it.