Updated Marker codebases to give control of the activation of the info window to the user of the plugin.
Added shouldShowInfoWindow to the JavaScript Marker object. This attribute can be modified with set, but has an explicit setter and getter. This attribute defaults to false. If this attribute is false, JavaScript Marker will never call the showInfoWindow native API. If it's true, calls to this API will be allowed assuming other pre-existing conditions allow.

All automated activation of the info window has been disabled in the native iOS and Android codebases.