fix(TP-2847): key cannot be nil crash
We were incorrectly using the removeObjectForKey API outside of the loop, thus
we were only removing the last key in the list.

In the case where there no keys in the objects map, a crash would occur since
key will be nil.

removeObjectForKey usage was moved inside the loop, and NSString* key declaration was also moved
inside the loop since it's not used outside of the loop at all.

The key value does change depending on a prefix, however this is a pattern used by all GMap plugin objects.
Therefore without fully understanding the rationale, I've kept this pattern.